Fehlen noch immer Details...

+++ cpVhostContainer/files/lib/acp/action/VhostContainerDeleteAction.class.php
@@ -13,12 +13,18 @@ class VhostContainerDeleteAction extends AbstractAction
 		parent :: execute();
 
 		$vhost = new VhostContainerEditor($_REQUEST['vhostContainerID']);
+		
+		$optionID = Domain :: getDomainOptionID('vhostContainerID');
+		
+		$sql = "SELECT COUNT(*) AS count
+				FROM cp" . CP_N . "domain_option_value 
+				WHERE domainOption" . $optionID . " = " . $vhost->vhostContainerID;
+		
+		$existCount = WCF :: getDB()->getFirstRow($sql);
 
-		if (WCF :: getUser()->getPermission('admin.cp.canDeleteVhostContainer'))
+		if (WCF :: getUser()->getPermission('admin.cp.canDeleteVhostContainer') && $existCount['count'] == 0)
 		{
 			$vhost->delete();
-			JobhandlerUtils :: addJob('deleteVhosts', 0, array('vhostContainerID' => $vhost->vhostContainerID), 'asap');
-			EventHandler :: fireAction($this, 'vhostContainerDeleted');
 		}
 		
 		$url = 'index.php?page=VhostContainerList&packageID='.PACKAGE_ID.SID_ARG_2ND_NOT_ENCODED;
